Refine the auron build for IDEA friendly
