+++ common/src/main/scala/org/apache/auron/util/SemanticVersion.scala
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.spark.sql.auron.util
+package org.apache.auron.util
 
 /**
  * Encapsulate a component version for the convenience of version checks.
