+++ pom.xml
@@ -30,9 +30,9 @@
   <packaging>pom</packaging>
 
   <modules>
+    <module>common</module>
     <module>spark-version-annotation-macros</module>
     <module>spark-extension</module>
-    <module>${shimPkg}</module>
     <module>hadoop-shim</module>
     <module>dev/mvn-build-helper/assembly</module>
     <module>dev/mvn-build-helper/proto</module>
@@ -56,6 +56,7 @@
     <maven.plugin.surefire.version>3.3.0</maven.plugin.surefire.version>
     <maven.plugin.shade.version>3.5.2</maven.plugin.shade.version>
     <os-maven-plugin.version>1.7.1</os-maven-plugin.version>
+    <maven.plugin.templating.version>1.0.0</maven.plugin.templating.version>
 
     <scalamacros.paradise.version>2.1.1</scalamacros.paradise.version>
     <semanticdb.version>4.8.1</semanticdb.version>
@@ -226,9 +227,6 @@
           <args>
             <arg>-Ywarn-unused</arg>
           </args>
-          <jvmArgs>
-            <jvmArg>-Dauron.shim=${shimName}</jvmArg>
-          </jvmArgs>
         </configuration>
         <dependencies>
           <dependency>
@@ -459,67 +457,79 @@
 
     <profile>
       <id>spark-3.0</id>
+      <modules>
+        <module>spark-extension-shims-spark3</module>
+      </modules>
       <properties>
-        <shimName>spark-3.0</shimName>
-        <shimPkg>spark-extension-shims-spark3</shimPkg>
         <scalaTestVersion>3.0.8</scalaTestVersion>
         <scalafmtVersion>3.0.0</scalafmtVersion>
         <sparkVersion>3.0.3</sparkVersion>
+        <sparkBinaryVersion>3.0</sparkBinaryVersion>
       </properties>
     </profile>
 
     <profile>
       <id>spark-3.1</id>
+      <modules>
+        <module>spark-extension-shims-spark3</module>
+      </modules>
       <properties>
-        <shimName>spark-3.1</shimName>
-        <shimPkg>spark-extension-shims-spark3</shimPkg>
         <scalaTestVersion>3.2.9</scalaTestVersion>
         <scalafmtVersion>3.0.0</scalafmtVersion>
         <sparkVersion>3.1.3</sparkVersion>
+        <sparkBinaryVersion>3.1</sparkBinaryVersion>
       </properties>
     </profile>
 
     <profile>
       <id>spark-3.2</id>
+      <modules>
+        <module>spark-extension-shims-spark3</module>
+      </modules>
       <properties>
-        <shimName>spark-3.2</shimName>
-        <shimPkg>spark-extension-shims-spark3</shimPkg>
         <scalaTestVersion>3.2.9</scalaTestVersion>
         <scalafmtVersion>3.0.0</scalafmtVersion>
         <sparkVersion>3.2.4</sparkVersion>
+        <sparkBinaryVersion>3.2</sparkBinaryVersion>
       </properties>
     </profile>
 
     <profile>
       <id>spark-3.3</id>
+      <modules>
+        <module>spark-extension-shims-spark3</module>
+      </modules>
       <properties>
-        <shimName>spark-3.3</shimName>
-        <shimPkg>spark-extension-shims-spark3</shimPkg>
         <scalaTestVersion>3.2.9</scalaTestVersion>
         <scalafmtVersion>3.0.0</scalafmtVersion>
         <sparkVersion>3.3.4</sparkVersion>
+        <sparkBinaryVersion>3.3</sparkBinaryVersion>
       </properties>
     </profile>
 
     <profile>
       <id>spark-3.4</id>
+      <modules>
+        <module>spark-extension-shims-spark3</module>
+      </modules>
       <properties>
-        <shimName>spark-3.4</shimName>
-        <shimPkg>spark-extension-shims-spark3</shimPkg>
         <scalaTestVersion>3.2.9</scalaTestVersion>
         <scalafmtVersion>3.0.0</scalafmtVersion>
         <sparkVersion>3.4.4</sparkVersion>
+        <sparkBinaryVersion>3.4</sparkBinaryVersion>
       </properties>
     </profile>
 
     <profile>
       <id>spark-3.5</id>
+      <modules>
+        <module>spark-extension-shims-spark3</module>
+      </modules>
       <properties>
-        <shimName>spark-3.5</shimName>
-        <shimPkg>spark-extension-shims-spark3</shimPkg>
         <scalaTestVersion>3.2.9</scalaTestVersion>
         <scalafmtVersion>3.0.0</scalafmtVersion>
         <sparkVersion>3.5.6</sparkVersion>
+        <sparkBinaryVersion>3.5</sparkBinaryVersion>
       </properties>
     </profile>
