+++ spark-extension-shims-spark3/pom.xml
@@ -31,6 +31,11 @@
   <dependencies>
     <dependency>
       <groupId>org.apache.auron</groupId>
+      <artifactId>auron-common_${scalaVersion}</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>org.apache.auron</groupId>
       <artifactId>spark-extension_${scalaVersion}</artifactId>
       <version>${project.version}</version>
     </dependency>
@@ -93,6 +98,12 @@
       <version>1.12.10</version>
     </dependency>
 
+    <dependency>
+      <groupId>org.apache.auron</groupId>
+      <artifactId>auron-common_${scalaVersion}</artifactId>
+      <version>${project.version}</version>
+      <type>test-jar</type>
+    </dependency>
     <dependency>
       <groupId>org.apache.spark</groupId>
       <artifactId>spark-core_${scalaVersion}</artifactId>
